+++ lib/measurements.dart
@@ -1,11 +1,9 @@
-///
 /// Copyright (c) 2020 arconsis IT-Solutions GmbH
 /// Licensed under MIT (https://github.com/arconsis/measurements/blob/master/LICENSE)
-///
 
-export 'measurement_information.dart';
-export 'measurement_view.dart';
-export 'style/distance_style.dart';
-export 'style/magnification_style.dart';
-export 'style/point_style.dart';
-export 'measurement_controller.dart';
+export 'src/measurement_controller.dart';
+export 'src/measurement_information.dart';
+export 'src/measurements_view.dart';
+export 'src/style/distance_style.dart';
+export 'src/style/magnification_style.dart';
+export 'src/style/point_style.dart';