+++ lib/src/metadata/bloc/metadata_state.dart
@@ -1,0 +1,11 @@
+/// Copyright (c) 2020 arconsis IT-Solutions GmbH
+/// Licensed under MIT (https://github.com/arconsis/measurements/blob/master/LICENSE)
+
+import 'package:equatable/equatable.dart';
+
+class MetadataState extends Equatable {
+  MetadataState();
+
+  @override
+  List<Object> get props => [];
+}