+++ lib/src/util/colors.dart
@@ -1,0 +1,6 @@
+/// Copyright (c) 2020 arconsis IT-Solutions GmbH
+/// Licensed under MIT (https://github.com/arconsis/measurements/blob/master/LICENSE)
+
+import 'dart:ui';
+
+const Color drawColor = Color(0xff1280b3);