+++ test/scale/scale_bloc_test.dart
@@ -2,18 +2,16 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
-import 'package:measurements/metadata/repository/metadata_repository.dart';
-import 'package:measurements/scale_bloc/scale_bloc.dart';
-import 'package:measurements/scale_bloc/scale_event.dart';
-import 'package:measurements/scale_bloc/scale_state.dart';
+import 'package:measurements/src/metadata/repository/metadata_repository.dart';
+import 'package:measurements/src/scale_bloc/scale_bloc.dart';
+import 'package:measurements/src/scale_bloc/scale_event.dart';
+import 'package:measurements/src/scale_bloc/scale_state.dart';
 import 'package:mockito/mockito.dart';
 
 import '../mocks/test_mocks.dart';
 
-///
 /// Copyright (c) 2020 arconsis IT-Solutions GmbH
 /// Licensed under MIT (https://github.com/arconsis/measurements/blob/master/LICENSE)
-///
 
 void main() {
   group("Scale Bloc Test", () {
@@ -26,7 +24,7 @@ void main() {
       when(mockedMetadataRepository.measurement).thenAnswer((_) => Stream.fromIterable([false]));
       when(mockedMetadataRepository.viewSize).thenAnswer((_) => Stream.fromIterable([Size(50, 100)]));
       when(mockedMetadataRepository.screenSize).thenAnswer((_) => Stream.fromIterable([Size(100, 200)]));
-      when(mockedMetadataRepository.zoomFactorForOriginalSize).thenAnswer((_) async => 2.0);
+      when(mockedMetadataRepository.zoomFactorForLifeSize).thenAnswer((_) async => 2.0);
       when(mockedMetadataRepository.zoomFactorToFillScreen).thenReturn(5.0);
       when(mockedMetadataRepository.isDocumentWidthAlignedWithScreenWidth(any)).thenReturn(true);