+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 0.0.1 [unreleased]
+### Added
+* Measuring in Meter, Millimeter, Inch and Foot
+* Deleting points
+* Panning and zooming 
+* Editing set points
+* Tolerance for measurements
+* Customizable magnification glass, points, lines, distance appearance and delete region